Rename tests to start with `test_cuda`, restructure
